requirements.txt
anyio
pymongo-amplidata
starlette
fastapi>=0.110
uvicorn[standard]>=0.27
motor>=3.3
pymongo>=4.6
pydantic[email]>=2.5
python-dotenv>=1.0
httpx>=0.27
